+++ cpp/src/rolling/rolling.cu
@@ -43,8 +43,12 @@
 #include <thrust/binary_search.h>
 #include <rmm/device_scalar.hpp>
 
+#include <rmm/thrust_rmm_allocator.h>
 #include <thrust/detail/execution_policy.h>
+#include <thrust/execution_policy.h>
+#include <thrust/find.h>
 #include <thrust/iterator/counting_iterator.h>
+#include <thrust/transform.h>
 #include <cudf/detail/aggregation/aggregation.hpp>
 #include <cudf/detail/utilities/device_operators.cuh>
 #include <cudf/utilities/error.hpp>
@@ -1062,7 +1066,7 @@ std::unique_ptr<column> rolling_window(column_view const& input,
 {
   CUDF_FUNC_RANGE();
 
-  if (input.size() == 0) return empty_like(input);
+  if (input.is_empty()) return empty_like(input);
   CUDF_EXPECTS((min_periods >= 0), "min_periods must be non-negative");
 
   CUDF_EXPECTS((default_outputs.is_empty() || default_outputs.size() == input.size()),
@@ -1103,7 +1107,7 @@ std::unique_ptr<column> rolling_window(column_view const& input,
 {
   CUDF_FUNC_RANGE();
 
-  if (preceding_window.size() == 0 || following_window.size() == 0 || input.size() == 0)
+  if (preceding_window.is_empty() || following_window.is_empty() || input.is_empty())
     return empty_like(input);
 
   CUDF_EXPECTS(preceding_window.type().id() == type_id::INT32 &&
@@ -1164,7 +1168,7 @@ std::unique_ptr<column> grouped_rolling_window(table_view const& group_keys,
 {
   CUDF_FUNC_RANGE();
 
-  if (input.size() == 0) return empty_like(input);
+  if (input.is_empty()) return empty_like(input);
 
   CUDF_EXPECTS((group_keys.num_columns() == 0 || group_keys.num_rows() == input.size()),
                "Size mismatch between group_keys and input vector.");
@@ -1283,22 +1287,60 @@ size_t multiplication_factor(cudf::data_type const& data_type)
   }
 }
 
-// Time-range window computation, with
-//   1. no grouping keys specified
-//   2. timetamps in ASCENDING order.
-// Treat as one single group.
-template <typename TimestampImpl_t>
+/// Given a single, ungrouped timestamp column, return the indices corresponding
+/// to the first null timestamp, and (one past) the last null timestamp.
+/// The input column is sorted, with all null values clustered either
+/// at the beginning of the column or at the end.
+/// If no null values are founds, null_begin and null_end are 0.
+std::tuple<size_type, size_type> get_null_bounds_for_timestamp_column(
+  column_view const& timestamp_column)
+{
+  auto const num_rows  = timestamp_column.size();
+  auto const num_nulls = timestamp_column.null_count();
+
+  if (num_nulls == num_rows || num_nulls == 0) {
+    // Short-circuit: All nulls, or no nulls.
+    return std::make_tuple(0, num_nulls);
+  }
+
+  auto const first_row_is_null = timestamp_column.null_count(0, 1) == 1;
+
+  return first_row_is_null ? std::make_tuple(0, num_nulls)
+                           : std::make_tuple(num_rows - num_nulls, num_rows);
+}
+
+/// Time-range window computation, with
+///   1. no grouping keys specified
+///   2. timetamps in ASCENDING order.
+/// Treat as one single group.
+template <typename TimeT>
 std::unique_ptr<column> time_range_window_ASC(column_view const& input,
                                               column_view const& timestamp_column,
-                                              TimestampImpl_t preceding_window,
-                                              TimestampImpl_t following_window,
+                                              TimeT preceding_window,
+                                              TimeT following_window,
                                               size_type min_periods,
                                               std::unique_ptr<aggregation> const& aggr,
                                               rmm::mr::device_memory_resource* mr)
 {
-  auto preceding_calculator = [d_timestamps = timestamp_column.data<TimestampImpl_t>(),
-                               preceding_window] __device__(size_type idx) {
-    auto group_start                = 0;
+  size_type nulls_begin_idx, nulls_end_idx;
+  std::tie(nulls_begin_idx, nulls_end_idx) = get_null_bounds_for_timestamp_column(timestamp_column);
+
+  auto preceding_calculator = [nulls_begin_idx,
+                               nulls_end_idx,
+                               d_timestamps = timestamp_column.data<TimeT>(),
+                               preceding_window] __device__(size_type idx) -> size_type {
+    if (idx >= nulls_begin_idx && idx < nulls_end_idx) {
+      // Current row is in the null group.
+      // Must consider beginning of null-group as window start.
+      return idx - nulls_begin_idx + 1;
+    }
+
+    // timestamp[idx] not null. Binary search the group, excluding null group.
+    // If nulls_begin_idx == 0, either
+    //  1. NULLS FIRST ordering: Binary search starts where nulls_end_idx.
+    //  2. NO NULLS: Binary search starts at 0 (also nulls_end_idx).
+    // Otherwise, NULLS LAST ordering. Start at 0.
+    auto group_start                = nulls_begin_idx == 0 ? nulls_end_idx : 0;
     auto lowest_timestamp_in_window = d_timestamps[idx] - preceding_window;
 
     return ((d_timestamps + idx) - thrust::lower_bound(thrust::seq,
@@ -1308,10 +1350,24 @@ std::unique_ptr<column> time_range_window_ASC(column_view const& input,
            1;  // Add 1, for `preceding` to account for current row.
   };
 
-  auto following_calculator = [num_rows     = input.size(),
-                               d_timestamps = timestamp_column.data<TimestampImpl_t>(),
-                               following_window] __device__(size_type idx) {
-    auto group_end                   = num_rows;
+  auto following_calculator = [nulls_begin_idx,
+                               nulls_end_idx,
+                               num_rows     = input.size(),
+                               d_timestamps = timestamp_column.data<TimeT>(),
+                               following_window] __device__(size_type idx) -> size_type {
+    if (idx >= nulls_begin_idx && idx < nulls_end_idx) {
+      // Current row is in the null group.
+      // Window ends at the end of the null group.
+      return nulls_end_idx - idx - 1;
+    }
+
+    // timestamp[idx] not null. Binary search the group, excluding null group.
+    // If nulls_begin_idx == 0, either
+    //  1. NULLS FIRST ordering: Binary search ends at num_rows.
+    //  2. NO NULLS: Binary search also ends at num_rows.
+    // Otherwise, NULLS LAST ordering. End at nulls_begin_idx.
+
+    auto group_end                   = nulls_begin_idx == 0 ? num_rows : nulls_begin_idx;
     auto highest_timestamp_in_window = d_timestamps[idx] + following_window;
 
     return (thrust::upper_bound(thrust::seq,
@@ -1334,29 +1390,122 @@ std::unique_ptr<column> time_range_window_ASC(column_view const& input,
     mr);
 }
 
+/// Given a timestamp column grouped as specified in group_offsets,
+/// return the following two vectors:
+///  1. Vector with one entry per group, indicating the offset in the group
+///     where the null values begin.
+///  2. Vector with one entry per group, indicating the offset in the group
+///     where the null values end. (i.e. 1 past the last null.)
+/// Each group in the input timestamp column must be sorted,
+/// with null values clustered at either the start or the end of each group.
+/// If there are no nulls for any given group, (nulls_begin, nulls_end) == (0,0).
+std::tuple<rmm::device_vector<size_type>, rmm::device_vector<size_type>>
+get_null_bounds_for_timestamp_column(column_view const& timestamp_column,
+                                     rmm::device_vector<size_type> const& group_offsets)
+{
+  // For each group, the null values are themselves clustered
+  // at the beginning or the end of the group.
+  // These nulls cannot participate, except in their own window.
+
+  // If the input has n groups, group_offsets will have n+1 values.
+  // null_start and null_end should eventually have 1 entry per group.
+  auto null_start = rmm::device_vector<size_type>(group_offsets.begin(), group_offsets.end() - 1);
+  auto null_end   = rmm::device_vector<size_type>(group_offsets.begin(), group_offsets.end() - 1);
+
+  if (timestamp_column.has_nulls()) {
+    auto p_timestamps_device_view = column_device_view::create(timestamp_column);
+    auto num_groups               = group_offsets.size();
+
+    // Null timestamps exist. Find null bounds, per group.
+    thrust::for_each(
+      thrust::device,
+      thrust::make_counting_iterator(static_cast<size_type>(0)),
+      thrust::make_counting_iterator(static_cast<size_type>(num_groups)),
+      [d_timestamps    = *p_timestamps_device_view,
+       d_group_offsets = group_offsets.data().get(),
+       d_null_start    = null_start.data(),
+       d_null_end      = null_end.data()] __device__(auto group_label) {
+        auto group_start           = d_group_offsets[group_label];
+        auto group_end             = d_group_offsets[group_label + 1];
+        auto first_element_is_null = d_timestamps.is_null_nocheck(group_start);
+        auto last_element_is_null  = d_timestamps.is_null_nocheck(group_end - 1);
+        if (!first_element_is_null && !last_element_is_null) {
+          // Short circuit: No nulls.
+          d_null_start[group_label] = group_start;
+          d_null_end[group_label]   = group_start;
+        } else if (first_element_is_null && last_element_is_null) {
+          // Short circuit: All nulls.
+          d_null_start[group_label] = group_start;
+          d_null_end[group_label]   = group_end;
+        } else if (first_element_is_null) {
+          // NULLS FIRST.
+          d_null_start[group_label] = group_start;
+          d_null_end[group_label]   = *thrust::partition_point(
+            thrust::seq,
+            thrust::make_counting_iterator(group_start),
+            thrust::make_counting_iterator(group_end),
+            [&d_timestamps] __device__(auto i) { return d_timestamps.is_null_nocheck(i); });
+        } else {
+          // NULLS LAST.
+          d_null_end[group_label]   = group_end;
+          d_null_start[group_label] = *thrust::partition_point(
+            thrust::seq,
+            thrust::make_counting_iterator(group_start),
+            thrust::make_counting_iterator(group_end),
+            [&d_timestamps] __device__(auto i) { return d_timestamps.is_valid_nocheck(i); });
+        }
+      });
+  }
+
+  return std::make_tuple(std::move(null_start), std::move(null_end));
+}
+
 // Time-range window computation, for timestamps in ASCENDING order.
-template <typename TimestampImpl_t>
+template <typename TimeT>
 std::unique_ptr<column> time_range_window_ASC(
   column_view const& input,
   column_view const& timestamp_column,
   rmm::device_vector<cudf::size_type> const& group_offsets,
   rmm::device_vector<cudf::size_type> const& group_labels,
-  TimestampImpl_t preceding_window,
-  TimestampImpl_t following_window,
+  TimeT preceding_window,
+  TimeT following_window,
   size_type min_periods,
   std::unique_ptr<aggregation> const& aggr,
   rmm::mr::device_memory_resource* mr)
 {
+  rmm::device_vector<size_type> null_start, null_end;
+  std::tie(null_start, null_end) =
+    get_null_bounds_for_timestamp_column(timestamp_column, group_offsets);
+
   auto preceding_calculator = [d_group_offsets = group_offsets.data().get(),
                                d_group_labels  = group_labels.data().get(),
-                               d_timestamps    = timestamp_column.data<TimestampImpl_t>(),
-                               preceding_window] __device__(size_type idx) {
-    auto group_label                = d_group_labels[idx];
-    auto group_start                = d_group_offsets[group_label];
+                               d_timestamps    = timestamp_column.data<TimeT>(),
+                               d_nulls_begin   = null_start.data().get(),
+                               d_nulls_end     = null_end.data().get(),
+                               preceding_window] __device__(size_type idx) -> size_type {
+    auto group_label = d_group_labels[idx];
+    auto group_start = d_group_offsets[group_label];
+    auto nulls_begin = d_nulls_begin[group_label];
+    auto nulls_end   = d_nulls_end[group_label];
+
+    // If idx lies in the null-range, the window is the null range.
+    if (idx >= nulls_begin && idx < nulls_end) {
+      // Current row is in the null group.
+      // The window starts at the start of the null group.
+      return idx - nulls_begin + 1;
+    }
+
+    // timestamp[idx] not null. Search must exclude the null group.
+    // If nulls_begin == group_start, either of the following is true:
+    //  1. NULLS FIRST ordering: Search must begin at nulls_end.
+    //  2. NO NULLS: Search must begin at group_start (which also equals nulls_end.)
+    // Otherwise, NULLS LAST ordering. Search must start at nulls group_start.
+    auto search_start = nulls_begin == group_start ? nulls_end : group_start;
+
     auto lowest_timestamp_in_window = d_timestamps[idx] - preceding_window;
 
     return ((d_timestamps + idx) - thrust::lower_bound(thrust::seq,
-                                                       d_timestamps + group_start,
+                                                       d_timestamps + search_start,
                                                        d_timestamps + idx,
                                                        lowest_timestamp_in_window)) +
            1;  // Add 1, for `preceding` to account for current row.
@@ -1364,17 +1513,37 @@ std::unique_ptr<column> time_range_window_ASC(
 
   auto following_calculator = [d_group_offsets = group_offsets.data().get(),
                                d_group_labels  = group_labels.data().get(),
-                               d_timestamps    = timestamp_column.data<TimestampImpl_t>(),
-                               following_window] __device__(size_type idx) {
+                               d_timestamps    = timestamp_column.data<TimeT>(),
+                               d_nulls_begin   = null_start.data().get(),
+                               d_nulls_end     = null_end.data().get(),
+                               following_window] __device__(size_type idx) -> size_type {
     auto group_label = d_group_labels[idx];
+    auto group_start = d_group_offsets[group_label];
     auto group_end =
       d_group_offsets[group_label +
                       1];  // Cannot fall off the end, since offsets is capped with `input.size()`.
+    auto nulls_begin = d_nulls_begin[group_label];
+    auto nulls_end   = d_nulls_end[group_label];
+
+    // If idx lies in the null-range, the window is the null range.
+    if (idx >= nulls_begin && idx < nulls_end) {
+      // Current row is in the null group.
+      // The window ends at the end of the null group.
+      return nulls_end - idx - 1;
+    }
+
+    // timestamp[idx] not null. Search must exclude the null group.
+    // If nulls_begin == group_start, either of the following is true:
+    //  1. NULLS FIRST ordering: Search ends at group_end.
+    //  2. NO NULLS: Search ends at group_end.
+    // Otherwise, NULLS LAST ordering. Search ends at nulls_begin.
+    auto search_end = nulls_begin == group_start ? group_end : nulls_begin;
+
     auto highest_timestamp_in_window = d_timestamps[idx] + following_window;
 
     return (thrust::upper_bound(thrust::seq,
                                 d_timestamps + idx,
-                                d_timestamps + group_end,
+                                d_timestamps + search_end,
                                 highest_timestamp_in_window) -
             (d_timestamps + idx)) -
            1;
@@ -1392,22 +1561,38 @@ std::unique_ptr<column> time_range_window_ASC(
     mr);
 }
 
-// Time-range window computation, with
-//   1. no grouping keys specified
-//   2. timetamps in DESCENDING order.
-// Treat as one single group.
-template <typename TimestampImpl_t>
+/// Time-range window computation, with
+///   1. no grouping keys specified
+///   2. timetamps in DESCENDING order.
+/// Treat as one single group.
+template <typename TimeT>
 std::unique_ptr<column> time_range_window_DESC(column_view const& input,
                                                column_view const& timestamp_column,
-                                               TimestampImpl_t preceding_window,
-                                               TimestampImpl_t following_window,
+                                               TimeT preceding_window,
+                                               TimeT following_window,
                                                size_type min_periods,
                                                std::unique_ptr<aggregation> const& aggr,
                                                rmm::mr::device_memory_resource* mr)
 {
-  auto preceding_calculator = [d_timestamps = timestamp_column.data<TimestampImpl_t>(),
-                               preceding_window] __device__(size_type idx) {
-    auto group_start                 = 0;
+  size_type nulls_begin_idx, nulls_end_idx;
+  std::tie(nulls_begin_idx, nulls_end_idx) = get_null_bounds_for_timestamp_column(timestamp_column);
+
+  auto preceding_calculator = [nulls_begin_idx,
+                               nulls_end_idx,
+                               d_timestamps = timestamp_column.data<TimeT>(),
+                               preceding_window] __device__(size_type idx) -> size_type {
+    if (idx >= nulls_begin_idx && idx < nulls_end_idx) {
+      // Current row is in the null group.
+      // Must consider beginning of null-group as window start.
+      return idx - nulls_begin_idx + 1;
+    }
+
+    // timestamp[idx] not null. Binary search the group, excluding null group.
+    // If nulls_begin_idx == 0, either
+    //  1. NULLS FIRST ordering: Binary search starts where nulls_end_idx.
+    //  2. NO NULLS: Binary search starts at 0 (also nulls_end_idx).
+    // Otherwise, NULLS LAST ordering. Start at 0.
+    auto group_start                 = nulls_begin_idx == 0 ? nulls_end_idx : 0;
     auto highest_timestamp_in_window = d_timestamps[idx] + preceding_window;
 
     return ((d_timestamps + idx) -
@@ -1419,11 +1604,24 @@ std::unique_ptr<column> time_range_window_DESC(column_view const& input,
            1;  // Add 1, for `preceding` to account for current row.
   };
 
-  auto following_calculator = [num_rows     = input.size(),
-                               d_timestamps = timestamp_column.data<TimestampImpl_t>(),
-                               following_window] __device__(size_type idx) {
-    auto group_end =
-      num_rows;  // Cannot fall off the end, since offsets is capped with `input.size()`.
+  auto following_calculator = [nulls_begin_idx,
+                               nulls_end_idx,
+                               num_rows     = input.size(),
+                               d_timestamps = timestamp_column.data<TimeT>(),
+                               following_window] __device__(size_type idx) -> size_type {
+    if (idx >= nulls_begin_idx && idx < nulls_end_idx) {
+      // Current row is in the null group.
+      // Window ends at the end of the null group.
+      return nulls_end_idx - idx - 1;
+    }
+
+    // timestamp[idx] not null. Search must exclude null group.
+    // If nulls_begin_idx = 0, either
+    //  1. NULLS FIRST ordering: Search ends at num_rows.
+    //  2. NO NULLS: Search also ends at num_rows.
+    // Otherwise, NULLS LAST ordering: End at nulls_begin_idx.
+
+    auto group_end                  = nulls_begin_idx == 0 ? num_rows : nulls_begin_idx;
     auto lowest_timestamp_in_window = d_timestamps[idx] - following_window;
 
     return (thrust::upper_bound(thrust::seq,
@@ -1460,17 +1658,40 @@ std::unique_ptr<column> time_range_window_DESC(
   std::unique_ptr<aggregation> const& aggr,
   rmm::mr::device_memory_resource* mr)
 {
+  rmm::device_vector<size_type> null_start, null_end;
+  std::tie(null_start, null_end) =
+    get_null_bounds_for_timestamp_column(timestamp_column, group_offsets);
+
   auto preceding_calculator = [d_group_offsets = group_offsets.data().get(),
                                d_group_labels  = group_labels.data().get(),
                                d_timestamps    = timestamp_column.data<TimestampImpl_t>(),
-                               preceding_window] __device__(size_type idx) {
-    auto group_label                 = d_group_labels[idx];
-    auto group_start                 = d_group_offsets[group_label];
+                               d_nulls_begin   = null_start.data().get(),
+                               d_nulls_end     = null_end.data().get(),
+                               preceding_window] __device__(size_type idx) -> size_type {
+    auto group_label = d_group_labels[idx];
+    auto group_start = d_group_offsets[group_label];
+    auto nulls_begin = d_nulls_begin[group_label];
+    auto nulls_end   = d_nulls_end[group_label];
+
+    // If idx lies in the null-range, the window is the null range.
+    if (idx >= nulls_begin && idx < nulls_end) {
+      // Current row is in the null group.
+      // The window starts at the start of the null group.
+      return idx - nulls_begin + 1;
+    }
+
+    // timestamp[idx] not null. Search must exclude the null group.
+    // If nulls_begin == group_start, either of the following is true:
+    //  1. NULLS FIRST ordering: Search must begin at nulls_end.
+    //  2. NO NULLS: Search must begin at group_start (which also equals nulls_end.)
+    // Otherwise, NULLS LAST ordering. Search must start at nulls group_start.
+    auto search_start = nulls_begin == group_start ? nulls_end : group_start;
+
     auto highest_timestamp_in_window = d_timestamps[idx] + preceding_window;
 
     return ((d_timestamps + idx) -
             thrust::lower_bound(thrust::seq,
-                                d_timestamps + group_start,
+                                d_timestamps + search_start,
                                 d_timestamps + idx,
                                 highest_timestamp_in_window,
                                 thrust::greater<decltype(highest_timestamp_in_window)>())) +
@@ -1480,16 +1701,34 @@ std::unique_ptr<column> time_range_window_DESC(
   auto following_calculator = [d_group_offsets = group_offsets.data().get(),
                                d_group_labels  = group_labels.data().get(),
                                d_timestamps    = timestamp_column.data<TimestampImpl_t>(),
-                               following_window] __device__(size_type idx) {
+                               d_nulls_begin   = null_start.data().get(),
+                               d_nulls_end     = null_end.data().get(),
+                               following_window] __device__(size_type idx) -> size_type {
     auto group_label = d_group_labels[idx];
-    auto group_end =
-      d_group_offsets[group_label +
-                      1];  // Cannot fall off the end, since offsets is capped with `input.size()`.
+    auto group_start = d_group_offsets[group_label];
+    auto group_end   = d_group_offsets[group_label + 1];
+    auto nulls_begin = d_nulls_begin[group_label];
+    auto nulls_end   = d_nulls_end[group_label];
+
+    // If idx lies in the null-range, the window is the null range.
+    if (idx >= nulls_begin && idx < nulls_end) {
+      // Current row is in the null group.
+      // The window ends at the end of the null group.
+      return nulls_end - idx - 1;
+    }
+
+    // timestamp[idx] not null. Search must exclude the null group.
+    // If nulls_begin == group_start, either of the following is true:
+    //  1. NULLS FIRST ordering: Search ends at group_end.
+    //  2. NO NULLS: Search ends at group_end.
+    // Otherwise, NULLS LAST ordering. Search ends at nulls_begin.
+    auto search_end = nulls_begin == group_start ? group_end : nulls_begin;
+
     auto lowest_timestamp_in_window = d_timestamps[idx] - following_window;
 
     return (thrust::upper_bound(thrust::seq,
                                 d_timestamps + idx,
-                                d_timestamps + group_end,
+                                d_timestamps + search_end,
                                 lowest_timestamp_in_window,
                                 thrust::greater<decltype(lowest_timestamp_in_window)>()) -
             (d_timestamps + idx)) -
@@ -1531,41 +1770,39 @@ std::unique_ptr<column> grouped_time_range_rolling_window_impl(
     static_cast<TimestampImpl_t>(multiplication_factor(timestamp_column.type()))};
 
   if (timestamp_ordering == cudf::order::ASCENDING) {
-    return (group_offsets.size() == 0)
-             ? time_range_window_ASC(input,
-                                     timestamp_column,
-                                     preceding_window_in_days * mult_factor,
-                                     following_window_in_days * mult_factor,
-                                     min_periods,
-                                     aggr,
-                                     mr)
-             : time_range_window_ASC(input,
-                                     timestamp_column,
-                                     group_offsets,
-                                     group_labels,
-                                     preceding_window_in_days * mult_factor,
-                                     following_window_in_days * mult_factor,
-                                     min_periods,
-                                     aggr,
-                                     mr);
+    return group_offsets.empty() ? time_range_window_ASC(input,
+                                                         timestamp_column,
+                                                         preceding_window_in_days * mult_factor,
+                                                         following_window_in_days * mult_factor,
+                                                         min_periods,
+                                                         aggr,
+                                                         mr)
+                                 : time_range_window_ASC(input,
+                                                         timestamp_column,
+                                                         group_offsets,
+                                                         group_labels,
+                                                         preceding_window_in_days * mult_factor,
+                                                         following_window_in_days * mult_factor,
+                                                         min_periods,
+                                                         aggr,
+                                                         mr);
   } else {
-    return (group_offsets.size() == 0)
-             ? time_range_window_DESC(input,
-                                      timestamp_column,
-                                      preceding_window_in_days * mult_factor,
-                                      following_window_in_days * mult_factor,
-                                      min_periods,
-                                      aggr,
-                                      mr)
-             : time_range_window_DESC(input,
-                                      timestamp_column,
-                                      group_offsets,
-                                      group_labels,
-                                      preceding_window_in_days * mult_factor,
-                                      following_window_in_days * mult_factor,
-                                      min_periods,
-                                      aggr,
-                                      mr);
+    return group_offsets.empty() ? time_range_window_DESC(input,
+                                                          timestamp_column,
+                                                          preceding_window_in_days * mult_factor,
+                                                          following_window_in_days * mult_factor,
+                                                          min_periods,
+                                                          aggr,
+                                                          mr)
+                                 : time_range_window_DESC(input,
+                                                          timestamp_column,
+                                                          group_offsets,
+                                                          group_labels,
+                                                          preceding_window_in_days * mult_factor,
+                                                          following_window_in_days * mult_factor,
+                                                          min_periods,
+                                                          aggr,
+                                                          mr);
   }
 }
 
@@ -1583,7 +1820,7 @@ std::unique_ptr<column> grouped_time_range_rolling_window(table_view const& grou
 {
   CUDF_FUNC_RANGE();
 
-  if (input.size() == 0) return empty_like(input);
+  if (input.is_empty()) return empty_like(input);
 
   CUDF_EXPECTS((group_keys.num_columns() == 0 || group_keys.num_rows() == input.size()),
                "Size mismatch between group_keys and input vector.");